Centralize exceptions
Place all exceptions in one place since some of the exceptions can
be reused for the agent as well.
